+++ service/collector_test.go
@@ -25,6 +25,7 @@ import (
 	"strings"
 	"syscall"
 	"testing"
+	"time"
 
 	"github.com/prometheus/common/expfmt"
 	"github.com/stretchr/testify/assert"
@@ -84,14 +85,16 @@ func TestCollector_StartAsGoRoutine(t *testing.T) {
 		}
 	}()
 
-	assert.Equal(t, Starting, <-col.GetStateChannel())
-	assert.Equal(t, Running, <-col.GetStateChannel())
+	assert.Eventually(t, func() bool {
+		return Running == col.GetState()
+	}, time.Second*2, time.Millisecond*200)
 
 	col.Shutdown()
 	col.Shutdown()
 	<-colDone
-	assert.Equal(t, Closing, <-col.GetStateChannel())
-	assert.Equal(t, Closed, <-col.GetStateChannel())
+	assert.Eventually(t, func() bool {
+		return Closed == col.GetState()
+	}, time.Second*2, time.Millisecond*200)
 }
 
 func TestCollector_Start(t *testing.T) {
@@ -125,8 +128,9 @@ func TestCollector_Start(t *testing.T) {
 		assert.NoError(t, col.Run(context.Background()))
 	}()
 
-	assert.Equal(t, Starting, <-col.GetStateChannel())
-	assert.Equal(t, Running, <-col.GetStateChannel())
+	assert.Eventually(t, func() bool {
+		return Running == col.GetState()
+	}, time.Second*2, time.Millisecond*200)
 	assert.Equal(t, col.logger, col.GetLogger())
 	assert.True(t, loggingHookCalled)
 
@@ -145,8 +149,9 @@ func TestCollector_Start(t *testing.T) {
 
 	col.signalsChannel <- syscall.SIGTERM
 	<-colDone
-	assert.Equal(t, Closing, <-col.GetStateChannel())
-	assert.Equal(t, Closed, <-col.GetStateChannel())
+	assert.Eventually(t, func() bool {
+		return Closed == col.GetState()
+	}, time.Second*2, time.Millisecond*200)
 }
 
 type mockColTelemetry struct{}
@@ -181,12 +186,14 @@ func TestCollector_ReportError(t *testing.T) {
 		assert.EqualError(t, col.Run(context.Background()), "failed to shutdown collector telemetry: err1")
 	}()
 
-	assert.Equal(t, Starting, <-col.GetStateChannel())
-	assert.Equal(t, Running, <-col.GetStateChannel())
+	assert.Eventually(t, func() bool {
+		return Running == col.GetState()
+	}, time.Second*2, time.Millisecond*200)
 	col.service.ReportFatalError(errors.New("err2"))
 	<-colDone
-	assert.Equal(t, Closing, <-col.GetStateChannel())
-	assert.Equal(t, Closed, <-col.GetStateChannel())
+	assert.Eventually(t, func() bool {
+		return Closed == col.GetState()
+	}, time.Second*2, time.Millisecond*200)
 }
 
 func assertMetrics(t *testing.T, prefix string, metricsPort uint16, mandatoryLabels []string) {