+++ CHANGELOG.md
@@ -4,9 +4,23 @@
 
 - Remove `defaultcomponents` from core (#4087)
 
+## 🧰 Bug fixes 🧰
+
+- Fix handling of corrupted records by persistent buffer (experimental) (#4475)
+
+## v0.40.0 Beta
+
 ## 🛑 Breaking changes 🛑
 
+- Package `client` refactored (#4416) and auth data included in it (#4422). Final PR to be merged in the next release (#4423)
 - Remove `pdata.AttributeMap.InitFromMap` (#4429)
+- Updated configgrpc `ToDialOptions` to support passing providers to instrumentation library (#4451)
+- Make state information propagation non-blocking on the collector (#4460)
+
+## 💡 Enhancements 💡
+
+- Add semconv 1.7.0 and 1.8.0 (#4452)
+- Added `feature-gates` CLI flag for controlling feature gate state. (#4368)
 
 ## v0.39.0 Beta